Initial Commit
Initial Commit

In this commit, AC pro is made compatible with the Aelia currency switcher plugin. This plugin allows you to switch the currency on the front end. You can specify the exchange rate on the admin side.

So Addon will capture the selected currency of the customer. If the customer had abandoned the cart with Indian rupee than it is captured

Same captured currency will be displayed in the abandoned orders tab, new orders details page. Abandoned cart reminder emails.

If we open the cart/checkout page from the email then the currency should be the same as it was in the email, Also the price of the product need to be same.